Modify files in order to use in browser
- move node.js only modules
- use snarkjs instead of zkutil
- add an api to instantiate User class
- fix tests
- add deploy script
