+++ src/hub.ts
@@ -3,12 +3,8 @@ import { Keypair, PubKey, Signature } from "maci-crypto";
 import { AddressInfo } from "ws";
 
 import { logger } from "./logger";
-import { getCounterSignHashedData, signMsg, verifySignedMsg } from ".";
-import {
-  RequestFailed,
-  DatabaseCorrupted,
-  HubRegistryNotFound
-} from "./exceptions";
+import { getCounterSignHashedData, signMsg } from ".";
+import { DatabaseCorrupted, HubRegistryNotFound } from "./exceptions";
 import {
   JoinReq,
   JoinResp,
@@ -18,22 +14,22 @@ import {
   SearchMessage3,
   msgType
 } from "./serialization";
-import { TLV, Short } from "./smp/serialization";
+import { TLV } from "./smp/serialization";
 import { bigIntToNumber } from "./smp/utils";
 import {
-  BaseServer,
   connect,
   IIPRateLimiter,
   TokenBucketRateLimiter,
   TRateLimitParams,
   IWebSocketReadWriter
 } from "./websocket";
-import { TIMEOUT, MAXIMUM_TRIALS, TIMEOUT_LARGE } from "./configs";
+import { BaseServer } from "./server";
+import { TIMEOUT, TIMEOUT_LARGE } from "./configs";
 import { objToHubRegistry, THubRegistryObj } from "./dataProvider";
 import { SMPStateMachine } from "./smp";
 import { hashPointToScalar } from "./utils";
 import { IAtomicDB, MerkleProof } from "./interfaces";
-import { genProofOfSMP, TProof } from "./circuits";
+import { genProofOfSMP } from "./circuits";
 import { IDBMap, DBMap } from "./db";
 import { SMPState1, SMPState2 } from "./smp/state";
 import {
@@ -42,17 +38,9 @@ import {
   SMPMessage2Wire,
   SMPMessage3Wire
 } from "./smp/v4/serialization";
-import { BabyJubPoint } from "./smp/v4/babyJub";
 
 type TUserRegistry = { userSig: Signature; hubSig: Signature };
 type TIterItem = [PubKey, TUserRegistry];
-type TSMPResult = {
-  a3: BigInt;
-  pa: BabyJubPoint;
-  ph: BabyJubPoint;
-  rh: BabyJubPoint;
-  proofOfSMP: TProof;
-};
 
 interface IUserStore extends AsyncIterable<TIterItem> {
   get(pubkey: PubKey): Promise<TUserRegistry | undefined>;
@@ -71,7 +59,11 @@ export class UserStore implements IUserStore {
   maxKeyLength = Point.size * 2;
 
   constructor(db: IAtomicDB) {
-    this.mapStore = new DBMap<TUserRegistry>(USER_STORE_PREFIX, db, this.maxKeyLength);
+    this.mapStore = new DBMap<TUserRegistry>(
+      USER_STORE_PREFIX,
+      db,
+      this.maxKeyLength
+    );
   }
 
   async getLength() {
@@ -79,7 +71,6 @@ export class UserStore implements IUserStore {
   }
 
   async *[Symbol.asyncIterator]() {
-
     for await (const obj of this.mapStore) {
       const pubkey = this.decodePubkey(obj.key);
       yield [pubkey, obj.value] as TIterItem;
@@ -134,7 +125,11 @@ const REGISTRY_STORE_PREFIX = "blind-find-hub-registry";
 export class RegistryStore {
   private dbMap: IDBMap<THubRegistryWithProof>;
   constructor(private readonly adminAddress: BigInt, db: IAtomicDB) {
-    this.dbMap = new DBMap<THubRegistryWithProof>(REGISTRY_STORE_PREFIX, db, this.getRegistryKey().length);
+    this.dbMap = new DBMap<THubRegistryWithProof>(
+      REGISTRY_STORE_PREFIX,
+      db,
+      this.getRegistryKey().length
+    );
   }
 
   private getRegistryKey(): string {
@@ -348,97 +343,3 @@ export class HubServer extends BaseServer {
     await this.userStore.removeAll();
   }
 }
-
-export const sendJoinHubReq = async (
-  ip: string,
-  port: number,
-  userPubkey: PubKey,
-  userSig: Signature,
-  hubPubkey: PubKey,
-  timeout: number = TIMEOUT
-): Promise<Signature> => {
-  const rwtor = await connect(ip, port);
-  const joinReq = new JoinReq(userPubkey, userSig);
-  const req = new TLV(new Short(msgType.JoinReq), joinReq.serialize());
-  rwtor.write(req.serialize());
-  const respBytes = await rwtor.read(timeout);
-  const resp = JoinResp.deserialize(respBytes);
-  const hasedData = getCounterSignHashedData(userSig);
-  if (!verifySignedMsg(hasedData, resp.hubSig, hubPubkey)) {
-    throw new RequestFailed("hub signature is invalid");
-  }
-  return resp.hubSig;
-};
-
-export const sendSearchReq = async (
-  ip: string,
-  port: number,
-  target: PubKey,
-  timeoutSmall: number = TIMEOUT,
-  timeoutLarge: number = TIMEOUT_LARGE,
-  maximumTrial: number = MAXIMUM_TRIALS
-): Promise<TSMPResult | null> => {
-  const rwtor = await connect(ip, port);
-
-  const msg0 = new SearchMessage0();
-  const req = new TLV(new Short(msgType.SearchReq), msg0.serialize());
-  rwtor.write(req.serialize());
-
-  let smpRes: TSMPResult | null = null;
-  let numTrials = 0;
-
-  const secret = hashPointToScalar(target);
-
-  while (numTrials < maximumTrial) {
-    logger.debug(
-      `sendSearchReq: starting trial ${numTrials}, waiting for msg1 from the server`
-    );
-    const stateMachine = new SMPStateMachine(secret);
-    const a3 = (stateMachine.state as SMPState1).s3;
-    const msg1Bytes = await rwtor.read(timeoutSmall);
-    const msg1 = SearchMessage1.deserialize(msg1Bytes);
-    logger.debug("sendSearchReq: received msg1");
-    // Check if there is no more candidates.
-    if (msg1.isEnd) {
-      break;
-    }
-    if (msg1.smpMsg1 === undefined) {
-      throw new Error(
-        "this should never happen, constructor already handles it for us"
-      );
-    }
-    const msg2 = stateMachine.transit(msg1.smpMsg1);
-    if (msg2 === null) {
-      throw new Error("this should never happen");
-    }
-    logger.debug("sendSearchReq: sending msg2");
-    rwtor.write(msg2.serialize());
-    const msg3Bytes = await rwtor.read(timeoutLarge);
-    logger.debug("sendSearchReq: received msg3");
-    const msg3 = SearchMessage3.deserialize(msg3Bytes);
-    stateMachine.transit(msg3.smpMsg3);
-    if (!stateMachine.isFinished()) {
-      throw new RequestFailed(
-        "smp should have been finished. there must be something wrong"
-      );
-    }
-    if (stateMachine.getResult()) {
-      logger.debug(
-        `sendSearchReq: SMP has matched, target=${target} is found`
-      );
-      const pa = SMPMessage2Wire.fromTLV(msg2).pb;
-      const smpMsg3 = SMPMessage3Wire.fromTLV(msg3.smpMsg3);
-      const ph = smpMsg3.pa;
-      const rh = smpMsg3.ra;
-      smpRes = {
-        a3,
-        pa,
-        ph,
-        rh,
-        proofOfSMP: msg3.proof
-      };
-    }
-    numTrials++;
-  }
-  return smpRes;
-};
