+++ test/circuits/utils.ts
@@ -1,9 +1,10 @@
 import * as path from "path";
-import { compileAndLoadCircuit } from "../../src/circuits";
-
+const circom = require("circom");
 const circomFilesDir = path.join(__dirname, "circom");
 
 export const compileCircuit = async (circomFileName: string) => {
-  const filePath = path.join(circomFilesDir, circomFileName);
-  return await compileAndLoadCircuit(filePath);
+  const circuitPath = path.join(circomFilesDir, circomFileName);
+  const circuit = await circom.tester(path.join(circuitPath));
+  await circuit.loadSymbols();
+  return circuit;
 };
